+++ netharn/util/__init__.py
@@ -48,7 +48,7 @@ from .util_json import (LossyJSONEncoder, NumpyEncoder,
                         write_json,)
 from .util_misc import (FlatIndexer, SupressPrint, align, align_lines,
                         strip_ansi,)
-from .util_resources import (ensure_ulimit,)
+from .util_resources import (ensure_ulimit, resource_usage)
 from .util_slider import (SlidingWindow, Stitcher,)
 from .util_slider_dep import (SlidingIndexDataset, SlidingSlices,)
 from .util_subextreme import (argsubmax, argsubmaxima,)
@@ -167,7 +167,7 @@ __all__ = ['ArrayAPI', 'BASE_COLORS', 'BatchNormContext', 'Boxes',
            'subpixel_translate', 'torch_ravel_multi_index', 'trainable_layers',
            'uniform', 'uniform32', 'util_dataframe', 'walk_json',
            'warp_points', 'warp_tensor', 'wide_strides_1d', 'write_arr',
-           'write_h5arr', 'write_json', 'zopen']
+           'write_h5arr', 'write_json', 'zopen', 'resource_usage']
 # </AUTOGEN_INIT>