+++ CHANGELOG.md
@@ -4,7 +4,33 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.5.12 - Unreleased
+## Version 0.5.13 - Unreleased
+
+### Added
+
+* Add 
+
+* Add preference `"log_resources"` to debug RAM and resource usage
+
+* Add preference `"dump_tensorboard"` to completely disable tensorboard
+
+* Ported SubsetSampler, PatchedRandomSampler, and PatchedBatchSampler from
+  bioharn
+
+
+### Fixed
+
+* `nh.api.Optimizer.coerce` now correctly respects lr and other params for
+  torch-optimizer optimizers.
+
+* `MultiLayerPerceptronNd` has several changes / fixes
+    - (BREAKING) always set `bias=False` in the hidden layers, this is now fixed. 
+    - dropout no longer is added if set to None
+    - option to apply a non-linearity / normalization to the output layer
+    - `output_shape_for` and `receptive_field_for` are now implemented via `_analytic_forward`
+
+
+## Version 0.5.12 - Released 2020-12-07
 
 ### Fixed